requirements.txt
discord.py-self==2.0.1
slack_sdk>=3.33.0
python-dotenv>=1.0.0
